requirements.txt
absl-py==1.3.0
addict==2.4.0
affine==2.4.0
aiohttp==3.8.3
aiosignal==1.3.1
albumentations==1.3.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
audioread==3.0.0
autopep8 @ file:///opt/conda/conda-bld/autopep8_1650463822033/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4==4.11.2
bitsandbytes-cuda116==0.26.0.post2
bitstring==4.0.1
bleach==6.0.0
Bottleneck @ file:///opt/conda/conda-bld/bottleneck_1657175564434/work
brotlipy==0.7.0
caerus==0.1.8
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cmake==3.25.2
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.2
contourpy==1.0.7
cryptography @ file:///croot/cryptography_1673298753778/work
cycler==0.11.0
datasets==2.11.0
debugpy==1.6.6
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml==0.7.1
dill==0.3.6
docker==6.0.1
docker-pycreds==0.4.0
einops==0.6.0
et-xmlfile==1.1.0
exceptiongroup==1.1.0
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.8.0
findpeaks==2.4.4
fire==0.4.0
flatbuffers==23.1.4
FlattenList==0.1.4
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2022.11.0
gast==0.4.0
gather==2023.1.20.1
gcc7==0.0.7
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.51.1
h5py==3.7.0
huggingface-hub==0.11.1
humanfriendly==10.0
icecream==2.1.3
idna @ file:///croot/idna_1666125576474/work
imageio==2.25.0
incremental==22.10.0
iniconfig==1.1.1
ipykernel==6.21.2
ipython @ file:///croot/ipython_1670919316550/work
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi @ file:///tmp/build/80754af9/jedi_1644315229345/work
Jinja2==3.1.2
joblib @ file:///croot/joblib_1666298844297/work
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.3.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
keras==2.11.0
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
kiwisolver==1.4.4
kornia==0.6.9
libclang==15.0.6.1
librosa==0.9.2
lightning-utilities==0.3.0
llvmlite==0.39.1
loguru @ file:///home/builder/ci_310/loguru_1641920743476/work
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mistune==2.0.5
mkl-fft==1.3.1
mkl-random @ file:///home/builder/ci_310/mkl_random_1641843545607/work
mkl-service==2.4.0
mmcv==1.7.1
mpmath==1.3.0
multidict==6.0.3
multiprocess==0.70.14
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.0
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numexpr @ file:///croot/numexpr_1668713893690/work
numpy @ file:///croot/numpy_and_numpy_base_1668593735768/work
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-tensorrt==8.4.1.5
oauthlib==3.2.2
omegaconf==2.2.3
onnxruntime==1.14.1
opencv-python==4.6.0.66
opencv-python-headless==4.7.0.68
openpyxl==3.1.1
opt-einsum==3.3.0
optimum==1.7.3
packaging @ file:///croot/packaging_1671697413597/work
pandas==1.5.1
pandocfilters==1.5.0
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pathtools==0.1.2
peakdetect==1.1
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.4.0
platformdirs==3.0.0
pluggy==1.0.0
pooch==1.6.0
progressbar2==4.2.0
prometheus-client==0.16.0
prompt-toolkit @ file:///croot/prompt-toolkit_1672387306916/work
protobuf==3.20.2
psutil==5.9.4
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pulsectl==22.3.2
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle @ file:///croot/pycodestyle_1674267221883/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydub==0.25.1
Pygments==2.13.0
pymagsac==0.3
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
pytest==7.2.0
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-json-logger==2.0.7
python-utils==3.4.5
pytlsd==0.0.2
pytorch-lightning==1.8.3.post1
pytz @ file:///croot/pytz_1671697431263/work
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
qudida==0.0.4
rasterio==1.3.5.post1
regex==2022.10.31
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
requests-oauthlib==1.3.1
resampy==0.4.2
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-image==0.19.3
scikit-learn @ file:///croot/scikit-learn_1667587546862/work
scipy==1.10.0
seaborn==0.12.2
Send2Trash==1.8.0
sentencepiece==0.1.97
sentry-sdk==1.15.0
setproctitle==1.3.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
sniffio==1.3.0
snuggs==1.4.7
SoundCard==0.4.2
sounddevice==0.4.5
soundfile==0.11.0
soupsieve==2.4
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
sympy==1.11.1
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
tensorrt==8.5.2.2
termcolor==2.1.1
terminado==0.17.1
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
tifffile==2023.2.2
timm==0.6.12
tinycss2==1.2.1
tokenizers==0.13.2
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tomli==2.0.1
torch==1.13.1
torch-tensorrt==1.3.0
torchaudio==0.13.0
torchmetrics==0.11.0
torchvision==0.14.1
torchviz==0.0.2
tornado==6.2
tqdm @ file:///opt/conda/conda-bld/tqdm_1664392687731/work
traitlets @ file:///croot/traitlets_1671143879854/work
transformers==4.27.4
typing_extensions @ file:///croot/typing_extensions_1669924550328/work
uri-template==1.2.0
urllib3 @ file:///croot/urllib3_1673575502006/work
venusian==3.0.0
wandb==0.13.10
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.2
wget==3.2
widgetsnbextension==4.0.5
wrapt==1.14.1
xxhash==3.2.0
yacs==0.1.8
yapf==0.32.0
yarl==1.8.2
